fix: return empty catalog name
The JDBC connection would return the current database name as the
current catalog name, but that catalog name would not be returned
in the ResultSet returned by getCatalogs(). This discrepancy causes
some tools to ignore or misinterpret the catalog and schema structure
of a Cloud Spanner database. Specifically, it breaks the autocomplete
feature of DBeaver.