tests: fixes jdbc query options test
This test was failing, because we were no longer recreating the jdbc
connection from scratch. One of the tests, set an environment variable
specifying the optimizer version, which was permanently applied, since
the connection was not re-created.

Here we make sure to start / stop the static server before each test.