test: fix assert for testConnectWithCredentialsAndOAuthToken
testConnectWithCredentialsAndOAuthToken test was failing with following error -
JdbcDriverTest.testConnectWithCredentialsAndOAuthToken:135 expected to contain: Cannot specify both credentials and an OAuth token
but was : INVALID_ARGUMENT: Specify only one of credentialsUrl, encodedCredentials, credentialsProvider and OAuth token.

Changing assertThat to account for this change. The exception message was changed as part of googleapis/java-spanner#1869.