test: modify tests to use correct quote escaping
Some tests used invalid escaping of quotes in PostgreSQL queries. They
assumed that a quote in a quoted string could be escaped by prefixing
it with a backslash. This is however not allowed in PostgreSQL, and
instead the quote must be escaped by including the same quote twice in
the string.
The Spanner Java client library contained a bug that allowed the
backslash-escaped quotes to be accepted. This bug has been fixed in
6.25.7 and these tests therefore need to be modified to be able to
update to that version.

Fixes the build error in #898