fix: support DDL/ClientSideStatement through (Statement/PreparedStatement).executeUpdate API
DDL statements and ClientSideStatements which return no result are now supported through PreparedStatement.executeUpdate and Statement.executeUpdate APIs in ConnectionImpl and should return 0L in that case. Exceptions are thrown for cases where the statement is expected to return a ResultSet (for example, DML Returning statements, ClientSideStatements which return ResultSet).