chore: add copyOf method for result sets
Adds a method for creating a JDBC ResultSet that creates a copy of a
Cloud Spanner ResultSet. This type of ResultSet can be used for small
result sets that are unlikely to be closed by the client application,
such as a ResultSet that holds the generated keys of an update
statement. If an application requests the JDBC driver to return the
generated keys, and then never gets and closes the generated keys
ResultSet, the underlying session and/or result stream could be
leaked.