feat: support endpoint connection URL property (#1582)
Added support for an 'endpoint' connection URL property. This property can
be added to either the Connection URL or the Properties object that is used
to create a Connection, and can be used in combination with the
autoConfigEmulator connection property to connect to an emulator instance
that uses a randomly assigned port number.

See https://github.com/googleapis/java-spanner-jdbc/blob/main/src/test/java/com/google/cloud/spanner/jdbc/ConcurrentTransactionOnEmulatorTest.java
for an example on how to use this property.