perf: use multiplexed sessions (#1673)
* perf: use multiplexed sessions

Enable the use of multiplexed sessions by default for queries in
auto-commit mode. Multiplexed sessions can handle any number of
queries concurrently. This means that the JDBC driver does not
need to check out a session exclusively from the internal session
pool in order to execute a query. Instead, a single multiplexed
session is enough for all queries that are executed by all JDBC
connections that connect to the same Spanner database. This
allows a higher degree of parallelism to be achieved from a
single client machine.

Note that due to how the JDBC API is defined, each JDBC
connection can only execute one query at a time. If you for
example want to execute 1000 queries in parallel, then you
also need to create 1000 JDBC connections. Spanner JDBC
connection are however lightweight, as each JDBC connection
internally uses a pool of gRPC channels.

It is recommended to enable the use of virtual threads to
achieve the highest possible degree of parallelism with the
JDBC driver. This option can be set by adding useVirtualThreads=true
to the JDBC connection URL. Note that virtual threads are only
supported on Java 21 and higher.

* test: add test for multi-use read-only transaction