feat: auto_batch_dml connection property (#1787)
* feat: auto_batch_dml connection property

The `auto_batch_dml` connection property enables automatic batching of
DML statements, for example when using Hibernate. When this connection
variable has been enabled, a JDBC connection will automatically buffer
DML statements in memory, and send all buffered DML statements to
Spanner as one batch when a query is executed or when the transaction
is committed. This can significantly reduce the number of round-trips
to Spanner, especially when using an ORM like Hibernate that generates
many small single-row DML statements.

* test: add test for auto_batch_dml