Update README to include all supported ConnectionProperties
Without this, users may believe that the supported set of connection
properties is exhaustive in the README. In addition there could be
lead-time between new features being added and the JDBC itself.

It should be noted that the version linked is `main` branch rather than
the one actively used by the JDBC driver.