chore: update test cases and regenerate documentation
The Spanner client now also uses a multiplexed session to determine the database
dialect that is in use. This means that BatchCreateSessions is not being called
at all in many cases.

Also re-generates the connection properties documentation. The typo in the
description of the new connection property must be fixed in the Spanner
client, as the list here is generated based on the source in the client.