feat: a single multiplexed session is used for all operations (#2162)
The JDBC driver now uses a single multiplexed session for all operations. This applies to all operations in auto-commit mode, in read-only transactions, and in read/write transactions.

One multiplexed session can execute an unlimited number of operations concurrently.

Configuring minSessions and maxSessions in the connection URL is therefore no longer needed for workloads that execute a large number of concurrent transactions.