feat: add transaction_timeout connection property
Adds a transaction_timeout connection property. This property can be set either in the
JDBC connection URL or by executing a statement like `SET transaction_timeout = '5s';`

The transaction timeout is the total time that a read/write transaction may take from
start to end. If the transaction is not able to finish within the given timeout, then
it will fail with a DEADLINE_EXCEEDED error.